moved getResolver to ultis
